pyproject.toml
[tool.poetry]
name = "pyproject"
version = "0.1.0"
description = ""
authors = ["Dinislam <dinislammutesov81@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
drf-spectacular = "^0.27.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
